+++ CHANGELOG.md
@@ -15,7 +15,8 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Fixed
 
-- None
+- [#1003](https://github.com/airblade/paper_trail/pull/1003) - detect Rails
+  presence with `Rails.application` 
 
 ## 8.0.0 (2017-10-04)