+++ CHANGELOG.md
@@ -16,6 +16,20 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Fixed
 
+- None
+
+## 8.0.1 (2017-10-25)
+
+### Breaking Changes
+
+- None
+
+### Added
+
+- None
+
+### Fixed
+
 - [#1003](https://github.com/airblade/paper_trail/pull/1003) - Warn when PT
   cannot be loaded because rails is not loaded yet.